fix: Corregir tipos de datos en initialize-database.js para compatibilidad con Foreign Keys - Cambiar orders.id de int(11) a bigint(20) en initialize-database.js - Cambiar order_items.order_id de int(11) a bigint(20) en initialize-database.js - Esto asegura que todos los scripts de BD tengan tipos compatibles - Resuelve el error persistente de Foreign Key incompatibility

+++ backend/scripts/initialize-database.js
@@ -161,7 +161,7 @@ const initializeDatabase = async () => {
     // Tabla de órdenes
     await connection.query(`
       CREATE TABLE IF NOT EXISTS orders (
-        id int(11) NOT NULL AUTO_INCREMENT,
+        id bigint(20) NOT NULL AUTO_INCREMENT,
         order_number varchar(50) NOT NULL,
         customer_type enum('registered','guest') NOT NULL,
         user_id bigint(20) DEFAULT NULL,
@@ -194,7 +194,7 @@ const initializeDatabase = async () => {
     await connection.query(`
       CREATE TABLE IF NOT EXISTS order_items (
         id int(11) NOT NULL AUTO_INCREMENT,
-        order_id int(11) NOT NULL,
+        order_id bigint(20) NOT NULL,
         product_id bigint(20) NOT NULL,
         product_name varchar(255) NOT NULL,
         product_price decimal(10,2) NOT NULL,
